EPA Region 10 dbf utilities. Version 5.7 July 13, 1998
This package is a set of utilities for translating data between dBASE
dbf, ASCII, and ARC/INFO formats.  Most of the programs run either as
ARC commands or as system commands independent of ARC/INFO.

An install script is included that configures, builds, tests, and
installs the package.  Pre-compiled packages are available for several
systems.

These programs were developed for EPA use, but have been released as
free software.  Anyone can use them.